bank account features

+++ src/main/java/com/bugboo/BookShop/service/OrderService.java
@@ -3,6 +3,7 @@ package com.bugboo.BookShop.service;
 import com.bugboo.BookShop.domain.*;
 import com.bugboo.BookShop.domain.dto.request.RequestCheckOutDTO;
 import com.bugboo.BookShop.domain.key.OrderDetailsId;
+import com.bugboo.BookShop.repository.BankAccountRepository;
 import com.bugboo.BookShop.repository.CartDetailsRepository;
 import com.bugboo.BookShop.repository.CartRepository;
 import com.bugboo.BookShop.repository.OrderRepository;
@@ -13,17 +14,20 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 
+
 @Service
 public class OrderService {
     private final OrderRepository orderRepository;
     private final OrderDetailsService orderDetailsService;
     private final CartRepository cartRepository;
     private final CartDetailsRepository cartDetailsRepository;
-    public OrderService(OrderRepository orderRepository, OrderDetailsService orderDetailsService, CartRepository cartRepository, CartDetailsRepository cartDetailsRepository) {
+    private final BankAccountRepository bankAccountRepository;
+    public OrderService(OrderRepository orderRepository, OrderDetailsService orderDetailsService, CartRepository cartRepository, CartDetailsRepository cartDetailsRepository, BankAccountRepository bankAccountRepository) {
         this.orderRepository = orderRepository;
         this.orderDetailsService = orderDetailsService;
         this.cartRepository = cartRepository;
         this.cartDetailsRepository = cartDetailsRepository;
+        this.bankAccountRepository = bankAccountRepository;
     }
 
     @Transactional
@@ -50,6 +54,17 @@ public class OrderService {
         }).toList();
 
         int total = cartDetails.stream().mapToInt(cartDetail -> cartDetail.getProduct().getPrice() * cartDetail.getQuantity()).sum();
+
+        BankAccount bankAccount = bankAccountRepository.findById(Long.valueOf(requestCheckOutDTO.getBankAccount().getId())).orElseThrow(() -> new AppException("Bank account not found",400));
+        if (bankAccount.getUser().getId() != currentUser.getId()) {
+            throw new AppException("Bank account not belongs to user",403);
+        }
+        if (bankAccount.getBalance() < total) {
+            throw new AppException("Insufficient balance",400);
+        }
+        bankAccount.setBalance(bankAccount.getBalance() - total);
+        bankAccountRepository.save(bankAccount);
+
         order.setTotal(total);
         cartDetailsRepository.deleteByCart(cart);
         cart.setSum(0);
